adding a module that normalizes console and includes a check for settings.debug
this module is already listed in the dependencies in app.js Though it returns a reference to console, the primary window.console was modified which means that this module does not need to be included anywhere else. Calls to console.log, dir, error, et al. will continue to work as normal so long as settings.debug=true